+++ integration/defer-test.ts
@@ -601,6 +601,7 @@ test.describe("non-aborted", () => {
   test("works with critical JSON like data", async ({ page }) => {
     let response = await fixture.requestDocument("/");
     let html = await response.text();
+    console.log(html);
     let criticalHTML = html.slice(0, html.indexOf("</html>") + 7);
     expect(criticalHTML).toContain(ROOT_ID);
     expect(criticalHTML).toContain(INDEX_ID);
