+++ integration/navigation-state-test.ts
@@ -16,10 +16,9 @@ const STATES = {
 
 const IDLE_STATE = {
   state: "idle",
-  type: "idle",
 };
 
-test.describe("rendering", () => {
+test.describe("navigation states", () => {
   let fixture: Fixture;
   let appFixture: AppFixture;
 
@@ -33,34 +32,34 @@ test.describe("rendering", () => {
   test.beforeAll(async () => {
     fixture = await createFixture({
       config: {
-        future: { v2_routeConvention: true },
+        future: {
+          v2_normalizeFormMethod: true,
+        },
       },
       files: {
         "app/root.jsx": js`
           import { useMemo, useRef } from "react";
-          import { Outlet, Scripts, useTransition } from "@remix-run/react";
+          import { Outlet, Scripts, useNavigation } from "@remix-run/react";
           export default function() {
-            const transition = useTransition();
-
-            const transitionsRef = useRef();
-            const transitions = useMemo(() => {
-              const savedTransitions = transitionsRef.current || [];
-              savedTransitions.push(transition);
-              transitionsRef.current = savedTransitions;
-              return savedTransitions;
-            }, [transition]);
-
+            const navigation = useNavigation();
+            const navigationsRef = useRef();
+            const navigations = useMemo(() => {
+              const savedNavigations = navigationsRef.current || [];
+              savedNavigations.push(navigation);
+              navigationsRef.current = savedNavigations;
+              return savedNavigations;
+            }, [navigation]);
             return (
               <html lang="en">
                 <head><title>Test</title></head>
                 <body>
                   <Outlet />
-                    {transition.state != "idle" && (
+                    {navigation.state != "idle" && (
                       <p id="loading-indicator">Loading...</p>
                     )}
                   <p>
-                    <code id="transitions">
-                      {JSON.stringify(transitions, null, 2)}
+                    <code id="navigations">
+                      {JSON.stringify(navigations, null, 2)}
                     </code>
                   </p>
                   <Scripts />
@@ -69,8 +68,7 @@ test.describe("rendering", () => {
             );
           }
         `,
-
-        "app/routes/_index.jsx": js`
+        "app/routes/index.jsx": js`
           import { Form, Link, useFetcher } from "@remix-run/react";
           export function loader() { return null; }
           export default function() {
@@ -126,7 +124,6 @@ test.describe("rendering", () => {
             );
           }
         `,
-
         [`app/routes/${STATES.NORMAL_LOAD}.jsx`]: js`
           export default function() {
             return (
@@ -136,7 +133,6 @@ test.describe("rendering", () => {
             );
           }
         `,
-
         [`app/routes/${STATES.LOADING_REDIRECT}.jsx`]: js`
           import { redirect } from "@remix-run/node";
           export function loader() {
@@ -150,7 +146,6 @@ test.describe("rendering", () => {
             );
           }
         `,
-
         [`app/routes/${STATES.SUBMITTING_LOADER}.jsx`]: js`
           export default function() {
             return (
@@ -160,7 +155,6 @@ test.describe("rendering", () => {
             );
           }
         `,
-
         [`app/routes/${STATES.SUBMITTING_LOADER_REDIRECT}.jsx`]: js`
           import { redirect } from "@remix-run/node";
           export function loader() {
@@ -174,7 +168,6 @@ test.describe("rendering", () => {
             );
           }
         `,
-
         [`app/routes/${STATES.SUBMITTING_ACTION}.jsx`]: js`
           export function loader() { return null; }
           export function action() { return null; }
@@ -186,7 +179,6 @@ test.describe("rendering", () => {
             );
           }
         `,
-
         [`app/routes/${STATES.SUBMITTING_ACTION_REDIRECT}.jsx`]: js`
           import { redirect } from "@remix-run/node";
           export function action() {
@@ -200,7 +192,6 @@ test.describe("rendering", () => {
             );
           }
         `,
-
         [`app/routes/${STATES.FETCHER_REDIRECT}.jsx`]: js`
           import { redirect } from "@remix-run/node";
           export function action() {
@@ -217,21 +208,20 @@ test.describe("rendering", () => {
     appFixture.close();
   });
 
-  test("transitions to normal load (Loading)", async ({ page }) => {
+  test("normal load (Loading)", async ({ page }) => {
     let app = new PlaywrightFixture(appFixture, page);
     await app.goto("/", true);
     await app.clickLink(`/${STATES.NORMAL_LOAD}`);
     await page.waitForSelector(`#${STATES.NORMAL_LOAD}`);
     await page.waitForSelector("#loading-indicator", { state: "hidden" });
 
-    let transitionsCode = await app.getElement("#transitions");
-    let transitionsJson = transitionsCode.text();
-    let transitions = JSON.parse(transitionsJson);
-    expect(transitions).toEqual([
+    let navigationsCode = await app.getElement("#navigations");
+    let navigationsJson = navigationsCode.text();
+    let navigations = JSON.parse(navigationsJson);
+    expect(navigations).toEqual([
       IDLE_STATE,
       {
         state: "loading",
-        type: "normalLoad",
         location: {
           pathname: `/${STATES.NORMAL_LOAD}`,
           search: "",
@@ -244,20 +234,19 @@ test.describe("rendering", () => {
     ]);
   });
 
-  test("transitions to normal redirect (LoadingRedirect)", async ({ page }) => {
+  test("normal redirect (LoadingRedirect)", async ({ page }) => {
     let app = new PlaywrightFixture(appFixture, page);
     await app.goto("/", true);
     await app.clickLink(`/${STATES.LOADING_REDIRECT}`);
     await page.waitForURL(/\?redirected/);
     await page.waitForSelector("#loading-indicator", { state: "hidden" });
-    let transitionsCode = await app.getElement("#transitions");
-    let transitionsJson = transitionsCode.text();
-    let transitions = JSON.parse(transitionsJson);
-    expect(transitions).toEqual([
+    let navigationsCode = await app.getElement("#navigations");
+    let navigationsJson = navigationsCode.text();
+    let navigations = JSON.parse(navigationsJson);
+    expect(navigations).toEqual([
       IDLE_STATE,
       {
         state: "loading",
-        type: "normalLoad",
         location: {
           pathname: `/${STATES.LOADING_REDIRECT}`,
           search: "",
@@ -268,17 +257,12 @@ test.describe("rendering", () => {
       },
       {
         state: "loading",
-        type: "normalRedirect",
         location: {
           pathname: "/",
           search: "?redirected",
           hash: "",
           state: {
             _isRedirect: true,
-            // These were private API for transition manager that are no longer
-            // needed with the new router so OK to disappear
-            // setCookie: false,
-            // type: "loader",
           },
           key: expect.any(String),
         },
@@ -287,22 +271,19 @@ test.describe("rendering", () => {
     ]);
   });
 
-  test("transitions to loader submission (SubmittingLoader)", async ({
-    page,
-  }) => {
+  test("loader submission (SubmittingLoader)", async ({ page }) => {
     let app = new PlaywrightFixture(appFixture, page);
     await app.goto("/", true);
     await app.clickSubmitButton(`/${STATES.SUBMITTING_LOADER}`);
     await page.waitForSelector(`#${STATES.SUBMITTING_LOADER}`);
     await page.waitForSelector("#loading-indicator", { state: "hidden" });
-    let transitionsCode = await app.getElement("#transitions");
-    let transitionsJson = transitionsCode.text();
-    let transitions = JSON.parse(transitionsJson);
-    expect(transitions).toEqual([
+    let navigationsCode = await app.getElement("#navigations");
+    let navigationsJson = navigationsCode.text();
+    let navigations = JSON.parse(navigationsJson);
+    expect(navigations).toEqual([
       IDLE_STATE,
       {
-        state: "submitting",
-        type: "loaderSubmission",
+        state: "loading",
         location: {
           pathname: `/${STATES.SUBMITTING_LOADER}`,
           search: "?key=value",
@@ -310,22 +291,16 @@ test.describe("rendering", () => {
           state: null,
           key: expect.any(String),
         },
-        submission: {
-          // Note: This is a bug in Remix but we're going to keep it that way
-          // in useTransition (including the back-compat version) and it'll be
-          // fixed with useNavigation
-          action: `/${STATES.SUBMITTING_LOADER}?key=value`,
-          encType: "application/x-www-form-urlencoded",
-          method: "GET",
-          key: expect.any(String),
-          formData: expect.any(Object),
-        },
+        formMethod: "GET",
+        formAction: `/${STATES.SUBMITTING_LOADER}`,
+        formEncType: "application/x-www-form-urlencoded",
+        formData: expect.any(Object),
       },
       IDLE_STATE,
     ]);
   });
 
-  test("transitions to loader submission redirect (LoadingLoaderSubmissionRedirect)", async ({
+  test("loader submission redirect (LoadingLoaderSubmissionRedirect)", async ({
     page,
   }) => {
     let app = new PlaywrightFixture(appFixture, page);
@@ -333,14 +308,13 @@ test.describe("rendering", () => {
     await app.clickSubmitButton(`/${STATES.SUBMITTING_LOADER_REDIRECT}`);
     await page.waitForURL(/\?redirected/);
     await page.waitForSelector("#loading-indicator", { state: "hidden" });
-    let transitionsCode = await app.getElement("#transitions");
-    let transitionsJson = transitionsCode.text();
-    let transitions = JSON.parse(transitionsJson);
-    expect(transitions).toEqual([
+    let navigationsCode = await app.getElement("#navigations");
+    let navigationsJson = navigationsCode.text();
+    let navigations = JSON.parse(navigationsJson);
+    expect(navigations).toEqual([
       IDLE_STATE,
       {
-        state: "submitting",
-        type: "loaderSubmission",
+        state: "loading",
         location: {
           pathname: `/${STATES.SUBMITTING_LOADER_REDIRECT}`,
           search: "",
@@ -348,58 +322,44 @@ test.describe("rendering", () => {
           state: null,
           key: expect.any(String),
         },
-        submission: {
-          action: `/${STATES.SUBMITTING_LOADER_REDIRECT}`,
-          encType: "application/x-www-form-urlencoded",
-          method: "GET",
-          key: expect.any(String),
-          formData: expect.any(Object),
-        },
+        formMethod: "GET",
+        formAction: `/${STATES.SUBMITTING_LOADER_REDIRECT}`,
+        formEncType: "application/x-www-form-urlencoded",
+        formData: expect.any(Object),
       },
       {
         state: "loading",
-        type: "loaderSubmissionRedirect",
         location: {
           pathname: "/",
           search: "?redirected",
           hash: "",
           state: {
             _isRedirect: true,
-            // These were private API for transition manager that are no longer
-            // needed with the new router so OK to disappear
-            // setCookie: false,
-            // type: "loader",
           },
           key: expect.any(String),
         },
-        submission: {
-          action: `/${STATES.SUBMITTING_LOADER_REDIRECT}`,
-          encType: "application/x-www-form-urlencoded",
-          method: "GET",
-          key: expect.any(String),
-          formData: expect.any(Object),
-        },
+        formMethod: "GET",
+        formAction: `/${STATES.SUBMITTING_LOADER_REDIRECT}`,
+        formEncType: "application/x-www-form-urlencoded",
+        formData: expect.any(Object),
       },
       IDLE_STATE,
     ]);
   });
 
-  test("transitions to action submission (SubmittingAction)", async ({
-    page,
-  }) => {
+  test("action submission (SubmittingAction)", async ({ page }) => {
     let app = new PlaywrightFixture(appFixture, page);
     await app.goto("/", true);
     await app.clickSubmitButton(`/${STATES.SUBMITTING_ACTION}`);
     await page.waitForSelector(`#${STATES.SUBMITTING_ACTION}`);
     await page.waitForSelector("#loading-indicator", { state: "hidden" });
-    let transitionsCode = await app.getElement("#transitions");
-    let transitionsJson = transitionsCode.text();
-    let transitions = JSON.parse(transitionsJson);
-    expect(transitions).toEqual([
+    let navigationsCode = await app.getElement("#navigations");
+    let navigationsJson = navigationsCode.text();
+    let navigations = JSON.parse(navigationsJson);
+    expect(navigations).toEqual([
       IDLE_STATE,
       {
         state: "submitting",
-        type: "actionSubmission",
         location: {
           pathname: `/${STATES.SUBMITTING_ACTION}`,
           search: "",
@@ -407,17 +367,13 @@ test.describe("rendering", () => {
           state: null,
           key: expect.any(String),
         },
-        submission: {
-          action: `/${STATES.SUBMITTING_ACTION}`,
-          encType: "application/x-www-form-urlencoded",
-          method: "POST",
-          key: expect.any(String),
-          formData: expect.any(Object),
-        },
+        formMethod: "POST",
+        formAction: `/${STATES.SUBMITTING_ACTION}`,
+        formEncType: "application/x-www-form-urlencoded",
+        formData: expect.any(Object),
       },
       {
         state: "loading",
-        type: "actionReload",
         location: {
           pathname: `/${STATES.SUBMITTING_ACTION}`,
           search: "",
@@ -425,19 +381,16 @@ test.describe("rendering", () => {
           state: null,
           key: expect.any(String),
         },
-        submission: {
-          action: `/${STATES.SUBMITTING_ACTION}`,
-          encType: "application/x-www-form-urlencoded",
-          method: "POST",
-          key: expect.any(String),
-          formData: expect.any(Object),
-        },
+        formMethod: "POST",
+        formAction: `/${STATES.SUBMITTING_ACTION}`,
+        formEncType: "application/x-www-form-urlencoded",
+        formData: expect.any(Object),
       },
       IDLE_STATE,
     ]);
   });
 
-  test("transitions to action submission redirect (LoadingActionRedirect)", async ({
+  test("action submission redirect (LoadingActionRedirect)", async ({
     page,
   }) => {
     let app = new PlaywrightFixture(appFixture, page);
@@ -445,14 +398,13 @@ test.describe("rendering", () => {
     await app.clickSubmitButton(`/${STATES.SUBMITTING_ACTION_REDIRECT}`);
     await page.waitForURL(/\?redirected/);
     await page.waitForSelector("#loading-indicator", { state: "hidden" });
-    let transitionsCode = await app.getElement("#transitions");
-    let transitionsJson = transitionsCode.text();
-    let transitions = JSON.parse(transitionsJson);
-    expect(transitions).toEqual([
+    let navigationsCode = await app.getElement("#navigations");
+    let navigationsJson = navigationsCode.text();
+    let navigations = JSON.parse(navigationsJson);
+    expect(navigations).toEqual([
       IDLE_STATE,
       {
         state: "submitting",
-        type: "actionSubmission",
         location: {
           pathname: `/${STATES.SUBMITTING_ACTION_REDIRECT}`,
           search: "",
@@ -460,43 +412,32 @@ test.describe("rendering", () => {
           state: null,
           key: expect.any(String),
         },
-        submission: {
-          action: `/${STATES.SUBMITTING_ACTION_REDIRECT}`,
-          encType: "application/x-www-form-urlencoded",
-          method: "POST",
-          key: expect.any(String),
-          formData: expect.any(Object),
-        },
+        formMethod: "POST",
+        formAction: `/${STATES.SUBMITTING_ACTION_REDIRECT}`,
+        formEncType: "application/x-www-form-urlencoded",
+        formData: expect.any(Object),
       },
       {
         state: "loading",
-        type: "actionRedirect",
         location: {
           pathname: "/",
           search: "?redirected",
           hash: "",
           state: {
             _isRedirect: true,
-            // These were private API for transition manager that are no longer
-            // needed with the new router so OK to disappear
-            // setCookie: false,
-            // type: "loader",
           },
           key: expect.any(String),
         },
-        submission: {
-          action: `/${STATES.SUBMITTING_ACTION_REDIRECT}`,
-          encType: "application/x-www-form-urlencoded",
-          method: "POST",
-          key: expect.any(String),
-          formData: expect.any(Object),
-        },
+        formMethod: "POST",
+        formAction: `/${STATES.SUBMITTING_ACTION_REDIRECT}`,
+        formEncType: "application/x-www-form-urlencoded",
+        formData: expect.any(Object),
       },
       IDLE_STATE,
     ]);
   });
 
-  test("transitions to fetcher action submission redirect (LoadingFetchActionRedirect)", async ({
+  test("fetcher action submission redirect (LoadingFetchActionRedirect)", async ({
     page,
   }) => {
     let app = new PlaywrightFixture(appFixture, page);
@@ -504,25 +445,20 @@ test.describe("rendering", () => {
     await app.clickSubmitButton(`/${STATES.FETCHER_REDIRECT}`);
     await page.waitForURL(/\?redirected/);
     await page.waitForSelector("#loading-indicator", { state: "hidden" });
-    let transitionsCode = await app.getElement("#transitions");
-    let transitionsJson = transitionsCode.text();
-    let transitions = JSON.parse(transitionsJson);
-    expect(transitions).toEqual([
+    let navigationsCode = await app.getElement("#navigations");
+    let navigationsJson = navigationsCode.text();
+    let navigations = JSON.parse(navigationsJson);
+    expect(navigations).toEqual([
       IDLE_STATE,
       {
         state: "loading",
-        type: "fetchActionRedirect",
         location: {
           pathname: "/",
           search: "?redirected",
           hash: "",
           state: {
-            _isRedirect: true,
             _isFetchActionRedirect: true,
-            // These were private API for transition manager that are no longer
-            // needed with the new router so OK to disappear
-            // setCookie: false,
-            // type: "loader",
+            _isRedirect: true,
           },
           key: expect.any(String),
         },
