+++ packages/remix-dev/__tests__/fixtures/replace-remix-magic-imports/app/routes/posts/admin/new.tsx
@@ -1,5 +1,6 @@
-import type { ActionFunction } from "remix";
-import { Form, redirect, json, useActionData, useTransition } from "remix";
+import type { ActionFunction } from "@remix-run/node";
+import { redirect, json } from "@remix-run/node";
+import { Form, useActionData, useNavigation } from "@remix-run/react";
 import invariant from "tiny-invariant";
 import { createPost } from "~/models/post.server";
 
@@ -37,8 +38,8 @@ const inputClassName = `w-full rounded border border-gray-500 px-2 py-1 text-lg`
 export default function NewPost() {
   const errors = useActionData();
 
-  const transition = useTransition();
-  const isCreating = Boolean(transition.submission);
+  const navigation = useNavigation();
+  const isCreating = Boolean(navigation.formData);
 
   return (
     <Form method="post">
