+++ packages/remix-dev/codemod/replace-remix-magic-imports/utils/export.ts
@@ -216,7 +216,6 @@ const exportsByRenderer: Record<Renderer, Partial<ExportNames>> = {
       "useResolvedPath",
       "useSearchParams",
       "useSubmit",
-      "useTransition",
     ],
   },
 };
