+++ packages/remix-react/index.tsx
@@ -62,7 +62,6 @@ export {
   NavLink,
   PrefetchPageLinks,
   LiveReload,
-  useTransition,
   useFetcher,
   useFetchers,
   useLoaderData,
